Update links and bazaar naming
This updates the download links to the various OSes. For the
Red Hat based OSes it points at the COPR repo that contains downloads
for CentOS, Red Hat, and Fedora. Updates the openSUSE link as well
to point at the Open Build Service.

As a drive by it also removes "bazaar" from the source code link.